Data.Monoid.OneLiner

Description

Derived methods for Semigroup and Monoid, using Generics.OneLiner and GHC.Generics.

Can be used for any types (deriving Generic) made with a single constructor, where every field is an instance of Semigroup (or Monoid, depending on the function).

Also includes a newtype wrapper that imbues any such data type with instant Semigroup and Monoid instances, which can one day be used with DerivingVia syntax to derive instances automatically.

Newtype wrapper

newtype GMonoid a

If a is a data type with a single constructor whose fields are all instances of Semigroup, then GMonoid a has a Semigroup instance.

If a is a data type with a single constructor whose fields are all instances of Monoid, then GMonoid a has a Monoid instance.

Will one day be able to be used with DerivingVia syntax, to derive instances automatically.

Constructors

GMonoid

Fields

getGMonoid :: a

Generics-derived methods

Semigroup

gSemigroup :: forall a. (ADTRecord a, Constraints a Semigroup) => a -> a -> a

Semigroup append (<>) implemented by calling <> on the components.

Monoid

gMappend :: forall a. (ADTRecord a, Constraints a Monoid) => a -> a -> a

Monoid append (mappend) implemented by calling <> on the components.

gMempty :: forall a. (ADTRecord a, Constraints a Monoid) => a

Monoid identity (mempty) implemented by using mempty for all of the components.