Data.Ord.OneLiner

Description

Derived methods for Eq and Ord, using Generics.OneLiner and GHC.Generics.

Can be used for any types (deriving Generic) where every field is an instance of Eq (or Ord).

Also includes a newtype wrapper that imbues any such data type with instant Eq and Ord instances, which can one day be used with DerivingVia syntax to derive instances automatically.

Newtype wrapper

newtype GOrd a

If a is a data type whose fields are all instances of Eq, then GOrd a has a Eq instance.

If a is a data type whose fields are all instances of Ord, then GOrd a has a Ord instance.

Will one day be able to be used with DerivingVia syntax, to derive instances automatically.

Constructors

GOrd

Fields

getGOrd :: a

Generics-derived methods

Eq

gEquals :: forall a. (ADT a, Constraints a Eq) => a -> a -> Bool

== implemented by using == between all of the components, lexicographically. First compares constructors.

gNotEquals :: forall a. (ADT a, Constraints a Eq) => a -> a -> Bool

/= implemented by using /= between all of the components, lexicographically. First compares constructors.

Ord

gCompare :: forall a. (ADT a, Constraints a Ord) => a -> a -> Ordering

compare implemented by using compare between all of the components, lexicographically. First compares constructors.

gLTE :: forall a. (ADT a, Constraints a Ord) => a -> a -> Bool

<= implemented by using <= between all of the components. First compares constructors.

gLT :: forall a. (ADT a, Constraints a Ord) => a -> a -> Bool

< implemented by using < between all of the components. First compares constructors.

gGTE :: forall a. (ADT a, Constraints a Ord) => a -> a -> Bool

>= implemented by using >= between all of the components. First compares constructors.

gGT :: forall a. (ADT a, Constraints a Ord) => a -> a -> Bool

> implemented by using > between all of the components. First compares constructors.

gMax :: forall a. (ADT a, Constraints a Ord) => a -> a -> a

max implemented by using max between all of the components. First compares constructors. If two items are equal, returns the second.

gMin :: forall a. (ADT a, Constraints a Ord) => a -> a -> a

min implemented by using min between all of the components. First compares constructors. If two items are equal, returns the first.