Numeric.OneLiner

Description

Derived methods for numeric typeclasses.

Can be used for any types (deriving Generic) made with a single constructor, where every field is an instance of Num (or Fractional or Floating, depending on the function).

Also includes a newtype wrapper that imbues any such data type with an instant Num (and Fractional and Floating) instance.

See README for details on usage instructions and motivations.

Newtype wrapper

newtype GNum a

If a is a data type with a single constructor whose fields are all instances of Num, then GNum a has a Num instance.

If a is a data type with a single constructor whose fields are all instances of Fractional, then GNum a has a Fractional instance.

If a is a data type with a single constructor whose fields are all instances of Floating, then GNum a has a Floating instance.

Constructors

GNum

Fields

getGNum :: a

Generics-derived methods

All of these implement the appropriate functions by carrying them over every field of the data type

gPlus :: forall a. (ADTRecord a, Constraints a Num) => a -> a -> a

gMinus :: forall a. (ADTRecord a, Constraints a Num) => a -> a -> a

gTimes :: forall a. (ADTRecord a, Constraints a Num) => a -> a -> a

gNegate :: forall a. (ADTRecord a, Constraints a Num) => a -> a

gAbs :: forall a. (ADTRecord a, Constraints a Num) => a -> a

gSignum :: forall a. (ADTRecord a, Constraints a Num) => a -> a

gFromInteger :: forall a. (ADTRecord a, Constraints a Num) => Integer -> a

Fractional

gDivide :: forall a. (ADTRecord a, Constraints a Fractional) => a -> a -> a

gRecip :: forall a. (ADTRecord a, Constraints a Fractional) => a -> a

gFromRational :: forall a. (ADTRecord a, Constraints a Fractional) => Rational -> a

Floating

gPi :: forall a. (ADTRecord a, Constraints a Floating) => a

gExp :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gLog :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gSqrt :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gPower :: forall a. (ADTRecord a, Constraints a Floating) => a -> a -> a

gLogBase :: forall a. (ADTRecord a, Constraints a Floating) => a -> a -> a

gSin :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gCos :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gTan :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gAsin :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gAcos :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gAtan :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gSinh :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gCosh :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gTanh :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gAsinh :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gAcosh :: forall a. (ADTRecord a, Constraints a Floating) => a -> a

gAtanh :: forall a. (ADTRecord a, Constraints a Floating) => a -> a