System.Random.OneLiner

Description

Derived methods for Random, using Generics.OneLiner and GHC.Generics.

Can be used for any types (deriving Generic) made with a single constructor, where every field is an instance of Random.

Also includes a newtype wrapper that imbues any such data type with instant Random instances, which can one day be used with DerivingVia syntax to derive instances automatically.

Single constructor

Newtype wrapper

newtype GRandom a

If a is a data type with a single constructor whose fields are all instances of Random, then GRandom a has a Random instance.

Will one day be able to be used with DerivingVia syntax, to derive instances automatically.

Only works with data types with single constructors. If you need it to work with types of multiple constructors, consider GRandomSum.

Constructors

GRandom

Fields

getGRandom :: a

Generics-derived methods

gRandomR :: forall a g. (ADTRecord a, Constraints a Random, RandomGen g) => (a, a) -> g -> (a, g)

randomR implemented by sequencing randomR between all components

Requires the type to have only a single constructor.

gRandom :: forall a g. (ADTRecord a, Constraints a Random, RandomGen g) => g -> (a, g)

random implemented by sequencing random for all components.

Requires the type to have only a single constructor.

gRandomRs :: forall a g. (ADTRecord a, Constraints a Random, RandomGen g) => (a, a) -> g -> [a]

randomRs implemented by repeatedly calling gRandomR.

gRandoms :: forall a g. (ADTRecord a, Constraints a Random, RandomGen g) => g -> [a]

randoms implemented by repeatedly calling gRandom.

gRandomRIO :: forall a. (ADTRecord a, Constraints a Random) => (a, a) -> IO a

randomRIO implemented by calling gRandomR on the global seed.

gRandomIO :: forall a. (ADTRecord a, Constraints a Random) => IO a

randomIO implemented by calling gRandom on the global seed.

Multiple constructor

Newtype wrapper

newtype GRandomSum a

If a is a data type whose fields are all instances of Random, then GRandom a has a Random instance.

Will one day be able to be used with DerivingVia syntax, to derive instances automatically.

A version of GRandom that works for data types with multiple constructors. If your type has only one constructor, it might be more performant to use GRandom.

Note that the "ranged" variants are partial: if given a range of items made with different constructors, will be error!

Constructors

GRandomSum

Fields

getGRandomSum :: a

Generics-derived methods

gRandomRSum :: forall a g. (ADT a, Constraints a Random, RandomGen g) => (a, a) -> g -> (a, g)

randomR implemented by sequencing randomR between all components.

If given a range of items made with different constructors, will be error!

gRandomSum :: forall a g. (ADT a, Constraints a Random, RandomGen g) => g -> (a, g)

random implemented by selecting a random constructor and sequencing random for all components.

gRandomRSums :: forall a g. (ADT a, Constraints a Random, RandomGen g) => (a, a) -> g -> [a]

randomRs implemented by repeatedly calling gRandomRSum.

If given a range of items made with different constructors, will be error!

gRandomSums :: forall a g. (ADT a, Constraints a Random, RandomGen g) => g -> [a]

randoms implemented by repeatedly calling gRandomSum.

gRandomRIOSum :: forall a. (ADT a, Constraints a Random) => (a, a) -> IO a

randomRIO implemented by calling gRandomRSum on the global seed.

If given a range of items made with different constructors, will be error!

gRandomIOSum :: forall a. (ADT a, Constraints a Random) => IO a

randomIO implemented by calling gRandom on the global seed.